tools/content-pipeline/requirements.txt
pymupdf>=1.24.0
anthropic>=0.30.0
python-dotenv>=1.0.0
Pillow>=10.0.0
